add and use generic card component
