+++ src/components/keycloak-card/keycloak-card.css
@@ -1,0 +1,3 @@
+  .keycloak__keycloak-card__footer-label {
+    margin-left: var(--pf-global--spacer--lg);
+  }
